Rename feature files using controller-service-repository naming
