+++ server/src/features/game/test/gameService.test.ts
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import { GameModel } from 'server/features/game/gameSchema';
 import { mockGame } from 'server/test/mock-data/mockGame';
-import { saveGames } from 'server/features/game/gameService';
+import { saveGames } from 'server/features/game/gameRepository';
 
 let mongoServer: MongoMemoryServer;
