+++ server/src/features/player-assignment/utils/removeMovedGamesFromUsers.ts
@@ -2,8 +2,8 @@ import moment from 'moment';
 import { logger } from 'server/utils/logger';
 import { GameDoc } from 'server/typings/game.typings';
 import { Game } from 'shared/typings/models/game';
-import { findGames } from 'server/features/game/gameService';
-import { findUsers, updateUser } from 'server/features/user/userService';
+import { findGames } from 'server/features/game/gameRepository';
+import { findUsers, updateUser } from 'server/features/user/userRepository';
 
 export const removeMovedGamesFromUsers = async (
   updatedGames: readonly Game[]
