+++ server/src/features/player-assignment/utils/test/removeMovedGamesFromUsers.test.ts
@@ -6,7 +6,7 @@ import { GameModel } from 'server/features/game/gameSchema';
 import { mockUser, mockSignup } from 'server/test/mock-data/mockUser';
 import { mockGame, mockGame2 } from 'server/test/mock-data/mockGame';
 import { removeMovedGamesFromUsers } from 'server/features/player-assignment/utils/removeMovedGamesFromUsers';
-import { saveSignup, saveUser } from 'server/features/user/userService';
+import { saveSignup, saveUser } from 'server/features/user/userRepository';
 
 let mongoServer: MongoMemoryServer;
