+++ server/src/features/serial/serialRepository.ts
