+++ server/src/features/settings/signuptimeService.ts
@@ -1,10 +1,6 @@
-import { saveSignupTime } from 'server/features/settings/settingsService';
-import { logger } from 'server/utils/logger';
+import { saveSignupTime } from 'server/features/settings/settingsRepository';
 
-// Add open signup time to server settings
-export const postSignupTime = async (signupTime: string): Promise<unknown> => {
-  logger.info('API call: POST /api/signuptime');
-
+export const storeSignupTime = async (signupTime: string): Promise<unknown> => {
   try {
     const response = await saveSignupTime(signupTime);
     return {
